Identify transactions via T-Code if present
The T-Code is a five-character identifier for the type of transaction.
It's language-agnostic, which is nice for identification purposes. A
list of these codes can (at the time of writing) be found here:
https://developer.paypal.com/docs/reports/reference/tcodes/

if the T-Code isn't present in the exported CSV, we fall back to looking
at the type.