Allow configuring a categorization function
To categorize payments sent and received. The user can optionally pass a
function in the `categorize` parameter of their import configuration,
which takes in the raw row from the CSV file. If the function returns a
false value, nothing new happens. Otherwise, the return value is taken
to be a string, which will be used to balance the entry in the ledger.